+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
 * Optimize to only scan for and set `Toot.trendingLinks` once if possible
+* Make `Toot.imageAttachments`, `videoAttachments`, etc. computed once in `Toot` constructor instead of methods
 
 ### v0.20.2
 * Repair media attachments with URL arguments in `remoteUrl`